OPRi dataset
The Opri dataset we collect is derived from truck types in real-world highway scenarios. When a truck appears under different cameras, each camera captures only one frontal image of the truck. Meanwhile, the Opri dataset is captured in a 24-hour-a-day uninterrupted manner by a network of cameras in two cities. The significant changes in truck appearance and illumination make the Opri dataset more challenging for domain adaptation and domain generalization tasks. We collect a total of four million truck images, and 8 researchers are invited to clean the Opri dataset for 2 months. For now, the Opri dataset is only open source for 130K images, and more data will be released soon.